Change type of variable that contain money from Double to BigDecimal

+++ src/main/java/com/belykh/finalProj/dao/LotDAO.java
@@ -4,6 +4,7 @@ import com.belykh.finalProj.entity.dbo.LotDBO;
 import com.belykh.finalProj.entity.dbo.LotState;
 import com.belykh.finalProj.exception.DAOException;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -15,7 +16,7 @@ public interface LotDAO {
     List<LotDBO> findAllLots() throws DAOException;
     List<LotDBO> findAllLotsByState(LotState state) throws DAOException;
     List<LotDBO> findAllLotsByStateAndOwnerId(Long ownerId,LotState state) throws DAOException;
-    boolean changeBuyerAndPrice(Long id,Long userId,Double newPrice) throws DAOException;
+    boolean changeBuyerAndPrice(Long id, Long userId, BigDecimal newPrice) throws DAOException;
     boolean changeState(Long id,LotState state) throws DAOException;
     void checkUnpaidLots() throws DAOException;
     boolean addLot(LotDBO lotDBO) throws DAOException;
